AdventOfCodeDay07/src/nativeTest/kotlin/Day07Tests.kt
import kotlin.test.Test
import kotlin.test.assertEquals

class Day07Tests {
    @Test
    fun solveSample1Test() {
        val lines = """
            light red bags contain 1 bright white bag, 2 muted yellow bags.
            dark orange bags contain 3 bright white bags, 4 muted yellow bags.
            bright white bags contain 1 shiny gold bag.
            muted yellow bags contain 2 shiny gold bags, 9 faded blue bags.
            shiny gold bags contain 1 dark olive bag, 2 vibrant plum bags.
            dark olive bags contain 3 faded blue bags, 4 dotted black bags.
            vibrant plum bags contain 5 faded blue bags, 6 dotted black bags.
            faded blue bags contain no other bags.
            dotted black bags contain no other bags.
        """.trimIndent().split("\n")

        val day07 = Day07(lines)
        val actual = day07.solvePart1()
        val expected = 4
        assertEquals(expected, actual)
    }

    @Test
    fun solveSample2Test() {
        val lines = """
            light red bags contain 1 bright white bag, 2 muted yellow bags.
            dark orange bags contain 3 bright white bags, 4 muted yellow bags.
            bright white bags contain 1 shiny gold bag.
            muted yellow bags contain 2 shiny gold bags, 9 faded blue bags.
            shiny gold bags contain 1 dark olive bag, 2 vibrant plum bags.
            dark olive bags contain 3 faded blue bags, 4 dotted black bags.
            vibrant plum bags contain 5 faded blue bags, 6 dotted black bags.
            faded blue bags contain no other bags.
            dotted black bags contain no other bags.
        """.trimIndent().split("\n")

        val day07 = Day07(lines)
        val actual = day07.solvePart2()
        val expected = 32
        assertEquals(expected, actual)
    }
}